+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Add a link to the GitHub diff like
 
 ### Fixed
 - action `Archive-Project` did not work after switching from `7za.exe` to `7z.exe`
+- expansion of config values `AppAdornmentBaseDir` and `AppRegistryBaseDir`
 
 ## [0.7.2] - 2016-02-12
 [Full Changelog](https://github.com/mastersign/bench/compare/v0.7.1...v0.7.2)